[구현 & 코너케이스] 선택 1713
